requirements.txt
PySide6>=6.5.0
paho-mqtt>=1.6.1
pyperclip>=1.8.2
protobuf>=4.21.0
Pillow>=9.0.0
zstandard>=0.21.0
